requirements.txt
Django==1.11.17
djangorestframework==3.11.0
redis==3.4.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2

idna==2.9
pytz==2020.1

requests==2.23.0
rq==1.3.0
urllib3==1.25.9
Jinja2==2.11.2
packaging==20.3
six==1.14.0
uritemplate==3.0.1

pip~=20.0.2
wheel~=0.34.2
attrs~=19.3.0
jsonschema~=3.2.0
pyrsistent~=0.16.0
setuptools~=46.0.0